Move class_attribute impl into Sidekiq::Worker
