+++ lib/sidekiq/worker.rb
@@ -4,7 +4,6 @@ require 'sidekiq/core_ext'
 
 module Sidekiq
 
-
   ##
   # Include this module in your worker class and you can easily create
   # asynchronous jobs:
@@ -147,6 +146,62 @@ module Sidekiq
         end
 
         Sidekiq::Client.new(pool).push(item)
+      end
+
+      def class_attribute(*attrs)
+        instance_reader = true
+        instance_writer = true
+
+        attrs.each do |name|
+          singleton_class.instance_eval do
+            undef_method(name) if method_defined?(name) || private_method_defined?(name)
+          end
+          define_singleton_method(name) { nil }
+
+          ivar = "@#{name}"
+
+          singleton_class.instance_eval do
+            m = "#{name}="
+            undef_method(m) if method_defined?(m) || private_method_defined?(m)
+          end
+          define_singleton_method("#{name}=") do |val|
+            singleton_class.class_eval do
+              undef_method(name) if method_defined?(name) || private_method_defined?(name)
+              define_method(name) { val }
+            end
+
+            if singleton_class?
+              class_eval do
+                undef_method(name) if method_defined?(name) || private_method_defined?(name)
+                define_method(name) do
+                  if instance_variable_defined? ivar
+                    instance_variable_get ivar
+                  else
+                    singleton_class.send name
+                  end
+                end
+              end
+            end
+            val
+          end
+
+          if instance_reader
+            undef_method(name) if method_defined?(name) || private_method_defined?(name)
+            define_method(name) do
+              if instance_variable_defined?(ivar)
+                instance_variable_get ivar
+              else
+                self.class.public_send name
+              end
+            end
+          end
+
+          if instance_writer
+            m = "#{name}="
+            undef_method(m) if method_defined?(m) || private_method_defined?(m)
+            attr_writer name
+          end
+        end
       end
 
     end
